Resync with MAGFest fork

+++ uber/site_sections/group_admin.py
@@ -5,12 +5,12 @@ from pockets import readable_join
 from pockets.autolog import log
 from pytz import UTC
 from sqlalchemy import and_, or_
-from sqlalchemy.orm import subqueryload
+from sqlalchemy.orm import joinedload
 
 from uber.config import c
 from uber.decorators import ajax, all_renderable, csrf_protected, log_pageview, site_mappable
 from uber.errors import HTTPRedirect
-from uber.forms import group as group_forms, load_forms
+from uber.forms import attendee as attendee_forms, group as group_forms, load_forms
 from uber.models import Attendee, Email, Event, Group, GuestGroup, GuestMerch, PageViewTracking, Tracking, SignedDocument
 from uber.utils import check, convert_to_absolute_url, validate_model, add_opt, SignNowRequest
 from uber.payments import ReceiptManager
@@ -28,17 +28,21 @@ class Root:
         return ''
 
     def index(self, session, message='', show_all=None):
-        if show_all:
-            groups = session.viewable_groups().limit(c.ROW_LOAD_LIMIT)
-        else:
-            groups = session.viewable_groups().filter(~Group.status.in_([c.DECLINED, c.IMPORTED, c.CANCELLED])).limit(c.ROW_LOAD_LIMIT)
-        dealer_groups = [group for group in groups if group.is_dealer]
+        groups = session.viewable_groups()
+        dealer_groups = groups.filter(Group.is_dealer == True)
+        guest_groups = groups.join(Group.guest)
+
+        if not show_all:
+            dealer_groups = dealer_groups.filter(~Group.status.in_([c.DECLINED, c.IMPORTED, c.CANCELLED]))
+
         return {
             'message': message,
-            'groups': groups.all(),
+            'groups': groups.options(joinedload(Group.attendees), joinedload(Group.leader), joinedload(Group.active_receipt)),
+            'guest_groups': guest_groups.options(joinedload(Group.attendees), joinedload(Group.leader)),
             'guest_checklist_items': GuestGroup(group_type=c.GUEST).sorted_checklist_items,
             'band_checklist_items': GuestGroup(group_type=c.BAND).sorted_checklist_items,
-            'dealer_groups':      len(dealer_groups),
+            'num_dealer_groups': dealer_groups.count(),
+            'dealer_groups':      dealer_groups.options(joinedload(Group.attendees), joinedload(Group.leader), joinedload(Group.active_receipt)),
             'dealer_badges':      sum(g.badges for g in dealer_groups),
             'tables':            sum(g.tables for g in dealer_groups),
             'show_all': show_all,
@@ -88,17 +92,22 @@ class Root:
                 receipt_items = ReceiptManager.auto_update_receipt(group, session.get_receipt_by_model(group), params)
                 session.add_all(receipt_items)
         else:
-            group = Group()
+            group = Group(tables=1) if new_dealer else Group()
 
-        if group.is_dealer or 'new_dealer' in params:
+        if group.is_dealer:
             form_list = ['AdminTableInfo', 'ContactInfo']
         else:
             form_list = ['AdminGroupInfo']
 
-        forms = load_forms(params, group, group_forms, form_list)
-        for form in forms.values():
-            if hasattr(form, 'new_badge_type'):
+        if group.is_new:
+            form_list.append('LeaderInfo')
+
+        forms = load_forms(params, group, form_list)
+        for form_name, form in forms.items():
+            if hasattr(form, 'new_badge_type') and not params.get('new_badge_type'):
                 form['new_badge_type'].data = group.leader.badge_type if group.leader else c.ATTENDEE_BADGE
+            if hasattr(form, 'new_ribbons') and not params.get('new_ribbons'):
+                form['new_ribbons'].data = group.leader.ribbon_ints if group.leader else []
             form.populate_obj(group, is_admin=True)
 
         signnow_last_emailed = None
@@ -132,47 +141,43 @@ class Root:
         group_info_form = forms.get('group_info', forms.get('table_info'))
 
         if cherrypy.request.method == 'POST':
-            new_with_leader = any(params.get(info) for info in ['first_name', 'last_name', 'email'])
-            message = message or self._required_message(params, ['name'])
-            
-            if not message and group.is_new and (params.get('guest_group_type') or new_dealer or group.is_dealer):
-                message = self._required_message(params, ['first_name', 'last_name', 'email'])
+            session.add(group)
 
-            if not message:
-                session.add(group)
+            if group.is_new and group.guest_group_type:
+                group.auto_recalc = False
 
-                if group.is_new and params.get('guest_group_type'):
-                    group.auto_recalc = False
+            if group.is_new or group.badges != group_info_form.badges.data:
+                test_permissions = Attendee(badge_type=group.new_badge_type, ribbon=group.new_ribbons, paid=c.PAID_BY_GROUP)
+                new_badge_status = c.PENDING_STATUS if not session.admin_can_create_attendee(test_permissions) else c.NEW_STATUS
+                message = session.assign_badges(
+                    group,
+                    group_info_form.badges.data or int(bool(group.leader_first_name)),
+                    new_badge_type=group.new_badge_type,
+                    new_ribbon_type=group.new_ribbons,
+                    badge_status=new_badge_status,
+                    )
 
-                if group.is_new or group.badges != group_info_form.badges.data:
-                    new_ribbon = params.get('ribbon', c.BAND if params.get('guest_group_type') == str(c.BAND) else None)
-                    new_badge_type = params.get('new_badge_type', c.ATTENDEE_BADGE)
-                    test_permissions = Attendee(badge_type=new_badge_type, ribbon=new_ribbon, paid=c.PAID_BY_GROUP)
-                    new_badge_status = c.PENDING_STATUS if not session.admin_can_create_attendee(test_permissions) else c.NEW_STATUS
-                    message = session.assign_badges(
-                        group,
-                        int(params.get('badges', 0)) or int(new_with_leader),
-                        new_badge_type=new_badge_type,
-                        new_ribbon_type=new_ribbon,
-                        badge_status=new_badge_status,
-                        )
-
-            if not message and group.is_new and new_with_leader:
+            if not message and group.is_new and group.leader_first_name:
                 session.commit()
                 leader = group.leader = group.attendees[0]
-                leader.first_name = params.get('first_name')
-                leader.last_name = params.get('last_name')
-                leader.email = params.get('email')
                 leader.placeholder = True
-                message = check(leader)
-                if message:
+                leader.badge_type = group.new_badge_type
+                leader.ribbon_ints = group.new_ribbons
+                leader_params = {key[7:]: val for key, val in params.items() if key.startswith('leader_')}
+                leader_forms = load_forms(leader_params, leader, ['PersonalInfo'])
+                all_errors = validate_model(leader_forms, leader, Attendee(**leader.to_dict()), is_admin=True)
+                if all_errors:
                     session.delete(group)
                     session.commit()
+                    message = ' '.join([item for sublist in all_errors.values() for item in sublist])
+                else:
+                    forms['personal_info'] = leader_forms['personal_info']
+                    forms['personal_info'].populate_obj(leader)
 
             if not message:
-                if params.get('guest_group_type'):
+                if group.guest_group_type:
                     group.guest = group.guest or GuestGroup()
-                    group.guest.group_type = params.get('guest_group_type')
+                    group.guest.group_type = group.guest_group_type
                 
                 if group.is_new and group.is_dealer:
                     if group.status == c.APPROVED and group.amount_unpaid:
@@ -199,26 +204,27 @@ class Root:
             'forms': forms,
             'signnow_last_emailed': signnow_last_emailed,
             'signnow_signed': signnow_signed,
-            'guest_group_type': params.get('guest_group_type', ''),
-            'badges': params.get('badges', group.badges if group else 0),
-            'first_name': params.get('first_name', ''),
-            'last_name': params.get('last_name', ''),
-            'email': params.get('email', ''),
             'new_dealer': new_dealer,
         }
     
     @ajax
-    def validate_dealer(self, session, form_list=[], **params):
+    def validate_group(self, session, form_list=[], new_dealer='', **params):
         if params.get('id') in [None, '', 'None']:
             group = Group()
         else:
             group = session.group(params.get('id'))
 
         if not form_list:
-            form_list = ['ContactInfo', 'AdminTableInfo']
+            if group.is_dealer or new_dealer:
+                form_list = ['AdminTableInfo', 'ContactInfo']
+            else:
+                form_list = ['AdminGroupInfo']
+
+            if group.is_new:
+                form_list.append('LeaderInfo')
         elif isinstance(form_list, str):
             form_list = [form_list]
-        forms = load_forms(params, group, group_forms, form_list)
+        forms = load_forms(params, group, form_list, get_optional=False)
 
         all_errors = validate_model(forms, group, Group(**group.to_dict()), is_admin=True)
         if all_errors:
